load-generator/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2018.11.29       # via geventhttpclient, requests
chardet==3.0.4            # via requests
click==7.0                # via flask
configargparse==1.2.3     # via locust
flask-basicauth==0.2.0    # via locust
flask==1.1.2              # via flask-basicauth, locust
gevent==20.4.0            # via geventhttpclient, locust
geventhttpclient==1.4.2   # via locust
googleapis-common-protos==1.51.0  # via -r requirements.in
greenlet==0.4.15          # via gevent
grpcio==1.27.1            # via -r requirements.in
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
locust==1.0.3             # via -r requirements.in
markupsafe==1.1.0         # via jinja2
msgpack==1.0.0            # via locust
protobuf==3.11.3          # via googleapis-common-protos
psutil==5.7.0             # via locust
pyzmq==17.0.0             # via locust
requests==2.21.0          # via locust
six==1.12.0               # via geventhttpclient, grpcio, protobuf
urllib3==1.24.2           # via requests
werkzeug==1.0.1           # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
